% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffusivity.R
\name{diffusivity}
\alias{diffusivity}
\title{Calculate the summed diffusivity of an air bubble}
\usage{
diffusivity(mol_rad, temp_c = 37, dyn_vis)
}
\arguments{
\item{mol_rad}{molecular radius of bubble (use mol_radius function)}

\item{temp_c}{temperature in degrees celcius 37 by default}

\item{dyn_vis}{dynamic viscosity of fluid surrounding bubble (see fizzTulz::dyn_vis)}
}
\description{
Function to calculate the summed diffusivity of air bubble suspended in fluid 
from the known temperature and viscosity of the system as well as the molecular radii of the bubble. 
Helper function for bubble lifetime model
}
\details{
diffusivity {fizzTulz}
}
