% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyn_vis.R
\name{dynVis}
\alias{dynVis}
\title{Calculate dynamic fluid viscosity of blood}
\usage{
dynVis(hct = 0.467)
}
\arguments{
\item{hem}{hematocrit fraction 0.467 by default}
}
\description{
Function to calculate the dynamic fluid viscosity of blood in Pascal-second (Pa-s)
from the known hematocrit fraction. Helper function for bubble lifetime model
}
\details{
dyn_vis {fizzTulz}
}
