% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_filt.R
\name{eff_filt}
\alias{eff_filt}
\title{Filter simple effects from emmeans object}
\usage{
eff_filt(e1, p = 0.05)
}
\arguments{
\item{e1}{Emmeans output object}

\item{p}{P-value to filter significant contrasts 0.05 by default}
}
\value{
A dataframe of meaningful contrasts
}
\description{
Pulls contrasts from emmeans object containing an interaction of two or more categorical 
variables where there is only one factor difference
}
\details{
eff_filt {fizzTulz}
}
