% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep_dyn_vis.R
\name{ep_dyn_vis}
\alias{ep_dyn_vis}
\title{Calculate dynamic fluid viscosity of blood}
\usage{
ep_dyn_vis(hct = 0.467)
}
\arguments{
\item{hct}{hematocrit fraction 0.467 by default}
}
\description{
Function to calculate the dynamic fluid viscosity of blood in Pascal-second (Pa-s)
from the known hematocrit fraction. Helper function for bubble lifetime model
}
\details{
ep_dyn_vis {fizzTulz}
}
