% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep_henry_sol.R
\name{ep_henry_sol}
\alias{ep_henry_sol}
\title{Calculate Henry's solubility constant}
\usage{
ep_henry_sol(temp_c = 37, kh = 1699)
}
\arguments{
\item{temp_c}{temperature in degrees celcius 37 by default}

\item{kh}{Henry's Law constant (1699 for air in water) 1699 by default}
}
\description{
Function to calculate the Henry's solubility constant for an air mixture in a bubble in fluid 
from the known temperature and viscosity of the system as well as the molecular radii of the bubble. 
Helper function for bubble lifetime model
}
\details{
ep_henry_sol {fizzTulz}
}
