% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lv_elast.R
\name{lv_elast}
\alias{lv_elast}
\title{Non-Invasive Left Ventricular End-Systolic Elastance}
\usage{
lv_elast(rtonset, rtend, ef, dbp, sbp, sv)
}
\arguments{
\item{rtonset}{Time from R-wave onset to aortic opening (s)}

\item{rtend}{Time from R-wave onset to aortic closing (s)}

\item{ef}{Ejection fraction}

\item{dbp}{Brachial diastolic blood pressure (mmHg)}

\item{sbp}{Brachial systolic blood pressure (mmHg)}

\item{sv}{Stroke volume (ml)}
}
\description{
Calculates End Systolic Elastance from cardiovascular and echocardiographic variables
}
\details{
lv_elast {fizzTulz}
}
\references{
Chen, C. H., Fetics, B., Nevo, E., Rochitte, C. E., Chiou, K. R., Ding, P. A., … Kass, D. A. (2001). Noninvasive single-beat determination of left ventricular end-systolic elastance in humans. Journal of the American College of Cardiology, 38(7), 2028–2034. https://doi.org/10.1016/S0735-1097(01)01651-5
}
