% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mol_radius.R
\name{mol_radius}
\alias{mol_radius}
\title{Calculate the molecular radius of a bubble}
\usage{
mol_radius(fiO2 = 0.21, fiN2 = 0.7896, fiCO2 = 4e-04)
}
\arguments{
\item{fiO2}{fraction of O2 in bubble 0.21 by default}

\item{fiN2}{fraction of N2 in bubble 0.7896 by default}

\item{fiCO2}{fraction of CO2 in bubble 0.0004 by default}
}
\description{
Function to calculate the molecular radius of an air bubble suspended in fluid 
from the known fraction of O2, N2 and CO2 within. Uses known molecular radii values of O2, N2, CO2 
Helper function for bubble lifetime model
}
\details{
mol_radius {fizzTulz}
}
