% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plusMinus.R
\name{plusMinus}
\alias{plusMinus}
\title{fizzTools}
\usage{
plusMinus(df1, catCols, digs = 1, df2 = NA)
}
\arguments{
\item{df1}{Dataframe with mean columns or with mean and sd columns}

\item{catCols}{Vector of grouping columns}

\item{digs}{Number of significant digits 1 by default}

\item{df2}{Dataframe with sd columns NA by default}
}
\value{
Dataframe with mean +/- sd values in character form
}
\description{
Creates mean +/- sd table
}
\examples{
df1 <- mtcars \%>\% group_by(gear, carb) \%>\%
  summarise_if(is.numeric, .funs = list("mean" = mean, "sd" = sd))
  plusMinus(df1, catCols = c('gear', 'carb'))

}
