% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd_split.R
\name{sd_split}
\alias{sd_split}
\title{Split Standard Deviation (error) columns}
\usage{
sd_split(df)
}
\arguments{
\item{df}{dataframe with mean and sd columns. Columns without proper suffix 
("_mean" and "_sd" (or "_sem)) will be ignored.}
}
\description{
Function that splits standard deviation (error) values in a column into two 
new positive and negative columns based on the mean variable value. Used to produce 
upper or lower error bars on bar plot. Dataframe should contain  unique mean 
and sd columns with matching variable name prefix and differing only by suffix 
"_mean" or "_sd" (or "_sem"). Note: If there are any unmatched mean and sd pairs
the function will fail.
}
\details{
sd_split {fizzTulz}
}
