% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sev_trans.R
\name{sev_trans}
\alias{sev_trans}
\title{Severinghaus Calculation for Hb-O2 Dissociation}
\usage{
sev_trans(po2 = NA, spo2 = NA)
}
\arguments{
\item{po2}{Partial pressure of arterial blood oxygen (mmHg)}

\item{spo2}{Peripheral oxygen saturation (fraction or %)}
}
\description{
Transforms PO2 values to SpO2 values and vice versa
}
\details{
sev_trans {fizzTulz}
}
\examples{
sevTrans(po2 = 55)
sevTrans(spo2 = 75)
sevTrans(spo2 = 0.75)

}
\references{
Severinghaus JW. Simple, accurate equations for human blood O2 dissociation computations. J Appl Physiol Respir Environ Exerc Physiol. 1979;46(3):599‐602. doi:10.1152/jappl.1979.46.3.599
}
