% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silo_select_load.R
\name{silo_select_load}
\alias{silo_select_load}
\title{Combine and Label Silo Selected Output Files}
\usage{
silo_select_load(file.dir)
}
\arguments{
\item{file.dir}{full 'selectData' directory path}
}
\description{
Combines files from silo > output > selectData into a single data frame with labels
}
\details{
silo_select_load {fizzTulz}
}
\examples{
silo_select_load('./silo/output/selectData/')


}
