% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smart_load.R
\name{smart_load}
\alias{smart_load}
\title{Load and Install Packages}
\usage{
smart_load(pkgs)
}
\arguments{
\item{pkgs}{Vector of r package names in string form}
}
\description{
Checks for installed packages, installs and loads them
}
\details{
smart_load {fizzTulz}
}
\examples{
physTools::smartLoad("ggplot2")
physTools::smartLoad(c("ggplot2", "lme4"))

}
