% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smartLoad.R
\name{smartLoad}
\alias{smartLoad}
\title{Load and Install Packages}
\usage{
smartLoad(pkgs)
}
\arguments{
\item{pkgs}{Vector of r package names in string form}
}
\description{
Checks for installed packages, installs and loads them
}
\details{
smartLoad {fizztools}
}
\examples{
physTools::smartLoad("ggplot2")
physTools::smartLoad(c("ggplot2", "lme4"))

}
